+++ src/delfino/constants.py
@@ -1,13 +1,14 @@
 from enum import Enum
 from pathlib import Path
-from typing import Final, Tuple
+from typing import Final
 
 ENTRY_POINT: Final[str] = "delfino"
 PYPROJECT_TOML_FILENAME: Final[str] = "pyproject.toml"
-DEFAULT_LOCAL_COMMAND_FOLDERS: Final[Tuple[Path, ...]] = (Path("commands"),)
+DEFAULT_LOCAL_COMMAND_FOLDERS: Final[tuple[Path, ...]] = (Path("commands"),)
 
 
 class PackageManager(Enum):
     POETRY = "poetry"
     PIPENV = "pipenv"
+    UV = "uv"
     UNKNOWN = "unknown"